Rename job.py to deprecation_job.py
Rename deprecate_image to process_job.
